Add files via upload
This is a bundled project activity of capacity 7.676 MW (i.e. AC capacity, which is equivalent to 8.82145 MW of DC capacity) which is grid connected solar PV power generation project in the District Bhilwara, in the State of Rajasthan, in India. The purpose of this renew power is for captive consumption. The promoter of the project activity is ‘Sudiva Spinners Private Limited’.